# docs/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Energy Drink App API",
    "version": "1.0.0",
    "description": "API for the Energy Drink Calculator and Recipe Application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/gdpr/data": {
      "get": {
        "summary": "GET /gdpr/data",
        "description": "",
        "parameters": [],
        "requestBody": null,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "ageVerified": {
                          "type": "boolean"
                        },
                        "verificationTimestamp": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "dataRetentionDays": {
                          "type": "number"
                        },
                        "dataPurposes": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "dataRecipients": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "legalBasis": {
                          "type": "string"
                        }
                      }
                    },
                    "rights": {
                      "type": "object",
                      "properties": {
                        "access": {
                          "type": "boolean"
                        },
                        "rectification": {
                          "type": "boolean"
                        },
                        "erasure": {
                          "type": "boolean"
                        },
                        "restriction": {
                          "type": "boolean"
                        },
                        "portability": {
                          "type": "boolean"
                        },
                        "objection": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "summary": "DELETE /gdpr/data",
        "description": "",
        "parameters": [],
        "requestBody": null,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "csrfAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "age-verified"
      },
      "csrfAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
...
  }
}